# src/config/docs/swagger-documentation.yml
openapi: 3.0.3
info:
  title: Documentación API Enpatados
  description: |-
    Esta es la documentación de la API de Enpatados, un emprendimiento que vende indumentaria, hecho con NodeJS utilizando Express, Sequeilize como ORM de MySQL y TypeScript, entre otras tecnologias, siguiendo el paradigma POO, buenas practicas y una nomeclatura. Para ver a detalle la API mirar en el repositorio: [https://github.com/Leaeraso/enpatados-poo](https://github.com/Leaeraso/enpatados-poo).
  version: 1.0.0
paths:
...
  /product/:
    get:
      tags:
        - Products
      summary: Get products
      parameters:
        - $ref: "#/parameters/search"
        - $ref: "#/parameters/page"
        - $ref: "#/parameters/page_size"
      responses:
        200:
          description: Return a product array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        404:
          description: Products not found
        500:
          description: Internal server error
    post:
      tags:
        - Products
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        200: 
          description: Product created successfully
        400:
          description: Product already exists
        404:
          description: Category/Subcategory not found
        500:
          description: Internal server error
      security:
        - bearerAuth: []
  /product/{id}:
    get:
      tags:
        - Products
      summary: Get product
      parameters:
        - $ref: "#/parameters/id"
      responses:
        200: 
         description: Return the product
         content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        404:
          description: Product not found
        500:
          description: Internal server error
    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - $ref: "#/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        200:
          description: Return the new product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        404:
          description: Product not found
        500:
          description: Internal server error
      security:
        - bearerAuth: []
    delete:
      tags:
        - Products
      summary: Delete product
      parameters:
        - $ref: "#/parameters/id"
      responses:
        200:
          description: Product deleted successfully
        404:
          description: Product not found
        500:
          description: Internal server error
      security:
        - bearerAuth: []
  /category/:
    get:
      tags:
        - Categories
      summary: Get categories
      responses:
        200:
          description: Return a category array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Categories not found
        500:
          description: Internal server error
    post:
      tags:
        - Categories
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        200:
          description: Category created successfully
        400:
          description: Category already exists
        500:
          description: Internal server error
      security:
        - bearerAuth: []
  /category/{id}:
    get:
      tags:
        - Categories
      summary: Get category
      parameters:
        - $ref: "#/parameters/id"
      responses:
        200:
          description: Return the category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        404:
          description: Category not found
        500:
          description: Internal server error
    put:
      tags:
        - Categories
      summary: Update category
      parameters:
        - $ref: "#/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        200:
          description: Return the new category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        404:
          description: Category not found
        500:
          description: Internal server error
      security:
        - bearerAuth: []
    delete:
      tags:
        - Categories
      summary: Delete category
      parameters:
        - $ref: "#/parameters/id"
      responses:
        200:
          description: Category deleted successfully
        404:
          description: Category not found
        500:
          description: Internal server error
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ProductCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category_id:
          type: integer
        subcategory_id:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category_id:
          type: integer
        subcategory_id:
          type: integer
    Product:
      type: object
      properties:
        product_id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        stock:
          type: integer
        category_id:
          type: integer
        subcategory_id:
          type: integer
...
    CategoryCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
    Category:
      type: object
      properties:
        category_id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
parameters:
  id:
    name: id
    in: path
    required: true
    schema:
      type: string
...
  page:
    name: page
    in: query
    description: Page number
    schema:
      type: number
      example: 1
  page_size:
    name: page_size
    in: query
    description: Page size
    schema:
      type: number
      example: 10
  search:
    name: search
    in: query
    description: Search
    schema:
      type: string